requirements.txt
python-dotenv
yfinance
packaging
fastapi
uvicorn
groq
phidata
typer
sqlalchemy
pgvector
psycopg[binary]
pypdf
google-genai
streamlit
duckduckgo-search
crewai
crewai_tools
langchain_google_genai